None value support on most readers and writers.
The UInt types don't support it (so we can use them for bitfields or such),
everything else does.

The float types use a selected signaling NaN value, so they should be
completely backwards compatible. (Nothing in python will make a signaling
NaN as far as I know.)

The time types use an invalid (all 0) value, so they are completely
backwards compatible.

The int types use T_MIN as None, so they are not completely backwards
compatible. Fortunately noone is really using this yet.

The string type uses NUL (character value 0) alone on the line for None.
This is backwards compatible, since NULs used to break everything.