+GzNumber, ints up to +-(2**1007 - 1) and floats.
This is surprisingly cheap (for floats and integers that are not actually
large).

These hash the same as GzFloat64 for floats and the same as GzInt64 for
ints that fit in GzInt64, but differently than GzBits64 for ints that don't
fit in GzInt64 but do fit in GzBits64 (2**63 <= x < 2**64). I consider this
not worth fixing.

generic_hash always gives what GzNumber would give for all numbers.