[dataset] Dataset instances pickle as strings.
Having a dependency on the actual dataset being available to load a pickle
with these in it is probably unexpected. Especially since they print like
strings.

Granted it's probably unexpected that they lose their methods on pickling
too, but that's the kind of problem you find right away, so it's better.